+++ v1/personio.go
@@ -11,11 +11,32 @@ import (
 	"strings"
 	"time"
 
-	"github.com/giantswarm/microerror"
 	"github.com/golang/glog"
 )
 
-const baseUrl = "https://api.personio.de/v1"
+const DefaultBaseUrl = "https://api.personio.de/v1"
+
+// Error is an error with an associated status code
+type Error interface {
+	error
+	Status() int
+}
+
+// StatusError represents an error with an associated HTTP status code
+type StatusError struct {
+	Err  error
+	Code int
+}
+
+// Allows StatusError to satisfy the error interface
+func (s StatusError) Error() string {
+	return s.Err.Error()
+}
+
+// Status returns the contained HTTP status code
+func (s StatusError) Status() int {
+	return s.Code
+}
 
 // PersonioBool is a custom boolean that can be unmarshalled from 0/1 and false/true
 type PersonioBool bool
@@ -44,11 +65,6 @@ type resultBody struct {
 
 // Auth is the response body of /auth
 type Auth struct {
-	Success bool `json:"success"`
-	Error   struct {
-		Code    int    `json:"code,omitempty"`
-		Message string `json:"message,omitempty"`
-	} `json:"error,omitempty"`
 	Data struct {
 		Token string `json:"token"`
 	} `json:"data,omitempty"`
@@ -60,6 +76,112 @@ type Attribute struct {
 	Value       interface{} `json:"value"`
 	Type        string      `json:"type"`
 	UniversalId string      `json:"universal_id"`
+}
+
+// GetIntValue returns a pointer to the attributes value as an int64 or nil if no such value is available
+func (a *Attribute) GetIntValue() *int64 {
+	if a.Type == "integer" && a.Value != nil {
+		switch a.Value.(type) {
+		case float64:
+			value := int64(a.Value.(float64))
+			return &value
+		}
+	}
+	return nil
+}
+
+// GetFloatValue returns a pointer to the attributes value as an float64 or nil if no such value is available
+func (a *Attribute) GetFloatValue() *float64 {
+	if (a.Type == "integer" || a.Type == "decimal") && a.Value != nil {
+		switch a.Value.(type) {
+		case float64:
+			value := a.Value.(float64)
+			return &value
+		}
+	}
+	return nil
+}
+
+// GetStringValue returns a pointer to the attributes value as string or nil if no such value is available
+func (a *Attribute) GetStringValue() *string {
+	if (a.Type == "standard" || a.Type == "multiline") && a.Value != nil {
+		switch a.Value.(type) {
+		case string:
+			value := a.Value.(string)
+			return &value
+		}
+	}
+	return nil
+}
+
+// GetListValue returns a pointer to the attributes value as string slice or nil if no such value is available
+func (a *Attribute) GetListValue() *[]string {
+	if a.Type == "list" && a.Value != nil {
+		switch a.Value.(type) {
+		case string:
+			value := strings.FieldsFunc(a.Value.(string), func(char rune) bool { return char == ',' })
+			return &value
+		}
+	}
+	return nil
+}
+
+// GetTimeValue returns a pointer to the attributes value as time.Time or nil if no such value is available
+func (a *Attribute) GetTimeValue() *time.Time {
+	if a.Type == "date" && a.Value != nil {
+		switch a.Value.(type) {
+		case string:
+			value, err := time.Parse(time.RFC3339, a.Value.(string))
+			if err == nil {
+				return &value
+			}
+		case time.Time:
+			value := a.Value.(time.Time)
+			return &value
+		}
+	}
+	return nil
+}
+
+// AttributeContainer is something that has object attributes of the elaborate and/or dynamic kind
+type AttributeContainer struct {
+	Attributes map[string]Attribute `json:"attributes"`
+}
+
+// GetIntAttribute returns a pointer to the specified attributes value as int64 or nil
+func (ac *AttributeContainer) GetIntAttribute(key string) *int64 {
+	attr := ac.Attributes[key]
+	return attr.GetIntValue()
+}
+
+// GetFloatAttribute returns a pointer to the specified attributes value as float64 or nil
+func (ac *AttributeContainer) GetFloatAttribute(key string) *float64 {
+	attr := ac.Attributes[key]
+	return attr.GetFloatValue()
+}
+
+// GetStringAttribute returns a pointer to the specified attributes value as string or nil
+func (ac *AttributeContainer) GetStringAttribute(key string) *string {
+	attr := ac.Attributes[key]
+	return attr.GetStringValue()
+}
+
+// GetListAttribute returns a pointer to the specified attributes value as string slice or nil
+func (ac *AttributeContainer) GetListAttribute(key string) *[]string {
+	attr := ac.Attributes[key]
+	return attr.GetListValue()
+}
+
+// GetTimeAttribute returns a pointer to the specified attributes value as time.Time or nil
+func (ac *AttributeContainer) GetTimeAttribute(key string) *time.Time {
+	attr := ac.Attributes[key]
+	return attr.GetTimeValue()
+}
+
+// Employee is a single employee entry
+type Employee struct {
+	Type string `json:"type"`
+	AttributeContainer
 }
 
 // TimeOff is a single time-off entry
@@ -79,11 +201,8 @@ type TimeOff struct {
 			Category string `json:"category"`
 		} `json:"attributes"`
 	} `json:"time_off_type"`
-	Employee struct {
-		Type       string `json:"type"`
-		Attributes map[string]Attribute
-	} `json:"employee"`
-	CreatedBy   string `json:"created_by"`
+	Employee    Employee `json:"employee"`
+	CreatedBy   string   `json:"created_by"`
 	Certificate struct {
 		Status string `json:"status"`
 	} `json:"certificate"`
@@ -91,56 +210,71 @@ type TimeOff struct {
 	UpdatedAt time.Time `json:"updated_at"`
 }
 
-// timeOffResult is the response body of /company/time-offs
-type timeOffResult struct {
-	Success bool `json:"success"`
-	Error   struct {
-		Code    int    `json:"code,omitempty"`
-		Message string `json:"message,omitempty"`
-	} `json:"error,omitempty"`
-	Data []struct {
-		Type       string  `json:"type"`
-		Attributes TimeOff `json:"attributes"`
-	} `json:"data,omitempty"`
+// EmployeeResult is the response body of /company/employee/{{id}}
+type employeeResult struct {
+	Data Employee `json:"data"`
+}
+
+// EmployeeResult is the response body of /company/employee/{{id}}
+type employeesResult struct {
+	Data []Employee `json:"data"`
+}
+
+// timeOffContainer is the typed object returned for time-offs by Personio
+type timeOffContainer struct {
+	Type       string  `json:"type"`
+	Attributes TimeOff `json:"attributes"`
+}
+
+// timeOffsResult is the response body of /company/time-offs
+type timeOffsResult struct {
+	Data []timeOffContainer `json:"data"`
 }
 
 // Credentials is the secret to authenticate with the Personio API v1
 type Credentials struct {
-	ClientId     string `json:"clientId,omitempty"`
-	ClientSecret string `json:"clientSecret,omitempty"`
+	ClientId     string `json:"clientId"`
+	ClientSecret string `json:"clientSecret"`
 	AccessToken  string `json:"accessToken,omitempty"`
 }
 
 // Client is a Personio API v1 instance
 type Client struct {
-	ctx    context.Context
-	client http.Client
-	secret Credentials
+	ctx     context.Context
+	baseUrl string
+	client  http.Client
+	secret  Credentials
 }
 
 // NewClient creates a new Client instance with the specified Credentials
-func NewClient(ctx context.Context, secret Credentials) (*Client, error) {
+func NewClient(ctx context.Context, baseUrl string, secret Credentials) (*Client, error) {
+
+	if baseUrl == "" {
+		baseUrl = DefaultBaseUrl
+	}
+
 	return &Client{
-		ctx:    ctx,
-		client: http.Client{Timeout: time.Duration(10) * time.Second},
-		secret: secret,
+		ctx:     ctx,
+		baseUrl: baseUrl,
+		client:  http.Client{Timeout: time.Duration(10) * time.Second},
+		secret:  secret,
 	}, nil
 }
 
 // doRequest processes the specified request, optionally handling authentication
-func (personio *Client) doRequest(request *http.Request, useAuthorization bool) ([]byte, error) {
+func (personio *Client) doRequest(request *http.Request, useAuthentication bool) ([]byte, error) {
 
-	// authorize
-	if useAuthorization && personio.secret.AccessToken == "" {
-		auth, err := personio.Authorize(personio.secret.ClientId, personio.secret.ClientSecret)
+	// authenticate
+	if useAuthentication && personio.secret.AccessToken == "" {
+		token, err := personio.Authenticate(personio.secret.ClientId, personio.secret.ClientSecret)
 		if err != nil {
-			return nil, microerror.Mask(err)
+			return nil, err
 		}
 
-		personio.secret.AccessToken = auth.Data.Token
+		personio.secret.AccessToken = token
 	}
 
-	if useAuthorization && personio.secret.AccessToken != "" {
+	if useAuthentication && personio.secret.AccessToken != "" {
 		(*request).Header.Set("Authorization", "Bearer "+personio.secret.AccessToken)
 		personio.secret.AccessToken = "" // token consumed
 	}
@@ -161,17 +295,17 @@ func (personio *Client) doRequest(request *http.Request, useAuthorization bool) 
 		}
 	}
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return nil, err
 	}
 
 	defer func(Body io.ReadCloser) {
 		err := Body.Close()
 		if err != nil {
-			glog.Error("Failed to close response body: %s", err)
+			glog.Errorf("Failed to close response body: %s", err)
 		}
 	}(response.Body)
 
-	if useAuthorization {
+	if useAuthentication {
 		// cycle or reset accessToken
 		nextAuthorization := strings.Replace(response.Header.Get("authorization"), "Bearer ", "", 1)
 		if nextAuthorization != "" {
@@ -179,29 +313,33 @@ func (personio *Client) doRequest(request *http.Request, useAuthorization bool) 
 		}
 	}
 
+	if response.StatusCode < http.StatusOK || response.StatusCode >= http.StatusMultipleChoices {
+		return nil, StatusError{errors.New(response.Status), response.StatusCode}
+	}
+
 	var body []byte
 	body, err = io.ReadAll(response.Body)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return nil, err
 	}
 
 	return body, nil
 }
 
 // doRequestJson processes the specified request assuming JSON data is exchanged
-func (personio *Client) doRequestJson(request *http.Request, useAuthorization bool) ([]byte, error) {
+func (personio *Client) doRequestJson(request *http.Request, useAuthentication bool) ([]byte, error) {
 
 	request.Header.Set("Accept", "application/json")
 
-	body, err := personio.doRequest(request, useAuthorization)
+	body, err := personio.doRequest(request, useAuthentication)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return nil, err
 	}
 
 	var result resultBody
 	err = json.Unmarshal(body, &result)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return nil, err
 	}
 
 	if !result.Success {
@@ -211,16 +349,16 @@ func (personio *Client) doRequestJson(request *http.Request, useAuthorization bo
 	return body, nil
 }
 
-// Authorize fetches a new access token for the given clientId and clientSecret
-func (personio *Client) Authorize(clientId string, clientSecret string) (*Auth, error) {
+// Authenticate fetches a new access token for the given clientId and clientSecret
+func (personio *Client) Authenticate(clientId string, clientSecret string) (string, error) {
 
 	form := url.Values{}
 	form.Add("client_id", clientId)
 	form.Add("client_secret", clientSecret)
 
-	req, err := http.NewRequest(http.MethodPost, baseUrl+"/auth", strings.NewReader(form.Encode()))
+	req, err := http.NewRequest(http.MethodPost, personio.baseUrl+"/auth", strings.NewReader(form.Encode()))
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return "", err
 	}
 
 	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
@@ -228,45 +366,101 @@ func (personio *Client) Authorize(clientId string, clientSecret string) (*Auth, 
 	var body []byte
 	body, err = personio.doRequestJson(req, false)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return "", err
 	}
 
 	var auth Auth
 	err = json.Unmarshal(body, &auth)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return "", err
 	}
 
-	return &auth, nil
+	return auth.Data.Token, nil
 }
 
-// GetTimeOffs returns the time-offs for the specified start and end dates (inclusive)
-func (personio *Client) GetTimeOffs(start time.Time, end time.Time) ([]*TimeOff, error) {
+// GetEmployee fetches one or multiple employees.json by optional ID
+func (personio *Client) GetEmployee(id int64) (*Employee, error) {
 
-	req, err := http.NewRequest(http.MethodGet, baseUrl+"/company/time-offs", nil)
+	req, err := http.NewRequest(http.MethodGet, personio.baseUrl+fmt.Sprintf("/company/employees/%d", id), nil)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return nil, err
 	}
-
-	query := req.URL.Query()
-	query.Add("start_date", start.Format(time.RFC3339))
-	query.Add("end_date", end.Format(time.RFC3339))
 
 	body, err := personio.doRequestJson(req, true)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return nil, err
 	}
 
-	var timeOffResult timeOffResult
+	var employeeResult employeeResult
+	err = json.Unmarshal(body, &employeeResult)
+	if err != nil {
+		return nil, err
+	}
+
+	// unpack single Employee element
+	return &employeeResult.Data, nil
+}
+
+// GetEmployees returns all employees
+func (personio *Client) GetEmployees() ([]*Employee, error) {
+
+	req, err := http.NewRequest(http.MethodGet, personio.baseUrl+"/company/employees", nil)
+	if err != nil {
+		return nil, err
+	}
+
+	body, err := personio.doRequestJson(req, true)
+	if err != nil {
+		return nil, err
+	}
+
+	var employeesResult employeesResult
+	err = json.Unmarshal(body, &employeesResult)
+	if err != nil {
+		return nil, err
+	}
+
+	// unpack TimeOff elements
+	employees := make([]*Employee, len(employeesResult.Data))
+	for i := range employeesResult.Data {
+		employees[i] = &employeesResult.Data[i]
+	}
+
+	return employees, nil
+}
+
+// GetTimeOffs returns the time-offs matching the specified start and end dates (inclusive, ignored if zero)
+func (personio *Client) GetTimeOffs(start *time.Time, end *time.Time) ([]*TimeOff, error) {
+
+	req, err := http.NewRequest(http.MethodGet, personio.baseUrl+"/company/time-offs", nil)
+	if err != nil {
+		return nil, err
+	}
+
+	query := req.URL.Query()
+	if start != nil {
+		query.Add("start_date", start.Format(time.RFC3339))
+	}
+	if end != nil {
+		query.Add("end_date", end.Format(time.RFC3339))
+	}
+	req.URL.RawQuery = query.Encode()
+
+	body, err := personio.doRequestJson(req, true)
+	if err != nil {
+		return nil, err
+	}
+
+	var timeOffResult timeOffsResult
 	err = json.Unmarshal(body, &timeOffResult)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return nil, err
 	}
 
 	// unpack TimeOff elements
 	timeOffs := make([]*TimeOff, len(timeOffResult.Data))
-	for i, entry := range timeOffResult.Data {
-		timeOffs[i] = &entry.Attributes
+	for i := range timeOffResult.Data {
+		timeOffs[i] = &timeOffResult.Data[i].Attributes
 	}
 
 	return timeOffs, nil